Data
Created a siteData object and exported to models to act as a source of data that will populate the site.
Inside models fetch is used to fetch data from an API to merge with site data for additional content.
Current setup does not include timestamps for blog posts but will go with current date at the time of render just for placeholder dates for this project or will setup placeholder posts in siteData with time and date stamps for placeholder posts.